+++ src/plugins/index.ts
@@ -1,21 +1,19 @@
-import router from '../router';
-import i18n from './i18n';
-import {createPinia} from 'pinia';
+// Types
+import type { App } from 'vue'
+import { createPinia } from 'pinia'
+import router from '../router'
 /**
  * plugins/index.ts
  *
  * Automatically included in `./src/main.ts`
  */
-
-// Types
-import type { App } from 'vue'
-
+import i18n from './i18n'
 // Plugins
 import vuetify from './vuetify'
 
 export function registerPlugins (app: App) {
- app.use(vuetify)
- app.use(createPinia());
- app.use(i18n);
- app.use(router);
+  app.use(vuetify)
+  app.use(createPinia())
+  app.use(i18n)
+  app.use(router)
 }